requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pycryptodome==3.17
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
sslcommerz-lib==1.0
tzdata==2022.7
urllib3==1.26.14
wrapt==1.11.2
